// demos/field-service-dispatch/src/generated/models/Fsd_ticketsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Fsd_ticketsfsd_category = {
  1: 'HVAC',
  2: 'Electrical',
  3: 'Plumbing',
  4: 'Appliance',
  5: 'Network',
  6: 'General'
} as const;
export type Fsd_ticketsfsd_category = keyof typeof Fsd_ticketsfsd_category;
export const Fsd_ticketsfsd_urgency = {
  1: 'Low',
  2: 'Medium',
  3: 'High',
  4: 'Critical'
} as const;
export type Fsd_ticketsfsd_urgency = keyof typeof Fsd_ticketsfsd_urgency;
export const Fsd_ticketsfsd_status = {
  1: 'New',
  2: 'Assigned',
  3: 'EnRoute',
  4: 'OnSite',
  5: 'Completed',
  6: 'Cancelled'
} as const;
export type Fsd_ticketsfsd_status = keyof typeof Fsd_ticketsfsd_status;
export const Fsd_ticketsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Fsd_ticketsstatecode = keyof typeof Fsd_ticketsstatecode;
export const Fsd_ticketsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Fsd_ticketsstatuscode = keyof typeof Fsd_ticketsstatuscode;

export interface Fsd_ticketsBase {
  fsd_address?: string;
  fsd_assignedon?: string;
  fsd_category?: Fsd_ticketsfsd_category;
  fsd_city?: string;
  fsd_completedon?: string;
  fsd_customername?: string;
  fsd_customerphone?: string;
  fsd_description?: string;
  fsd_latitude?: number;
  fsd_longitude?: number;
  fsd_name?: string;
  fsd_priorityscore?: number;
  fsd_reportedon?: string;
  fsd_resolutionnotes?: string;
  fsd_scheduledfor?: string;
  fsd_slahours?: number;
  fsd_status?: Fsd_ticketsfsd_status;
  fsd_technicianemail?: string;
  fsd_technicianname?: string;
  fsd_ticketid: string;
  fsd_urgency?: Fsd_ticketsfsd_urgency;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Fsd_ticketsstatecode;
  statuscode?: Fsd_ticketsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Fsd_tickets extends Fsd_ticketsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  fsd_categoryname?: string;
  fsd_statusname?: string;
  fsd_urgencyname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}